# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PixelPeek steganography studio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: steganography
  - name: analysis
  - name: history
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register — sends OTP to email before creating account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "200":
          description: OTP sent — verify email to complete registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginOtpPending"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in — returns OTP challenge when credentials are valid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OTP required (2FA step)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginOtpPending"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send/resend an OTP for a given purpose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResult"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify an OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: OTP verified — returns auth token or reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOtpResult"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: OTP sent to email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordResult"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /encode:
    post:
      operationId: encodeFile
      tags: [steganography]
      summary: Hide a message in a carrier file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/EncodeInput"
      responses:
        "200":
          description: Encoded file download
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncodeResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /decode:
    post:
      operationId: decodeFile
      tags: [steganography]
      summary: Reveal a hidden message from a carrier file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DecodeInput"
      responses:
        "200":
          description: Decoded message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecodeResult"
        "400":
          description: No message found or wrong key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /detect:
    post:
      operationId: detectSteganography
      tags: [analysis]
      summary: Analyze a file for hidden content
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DetectInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectResult"
        "400":
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [history]
      summary: Get operation statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /events:
    get:
      operationId: listEvents
      tags: [history]
      summary: List recent operations
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    RegisterBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    LoginOtpPending:
      type: object
      properties:
        requiresOtp:
          type: boolean
        pendingToken:
          type: string
      required: [requiresOtp, pendingToken]

    SendOtpBody:
      type: object
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          enum: [login, forgot-password, register]
      required: [email, purpose]

    SendOtpResult:
      type: object
      properties:
        message:
          type: string
      required: [message]

    VerifyOtpBody:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
          minLength: 6
          maxLength: 6
        purpose:
          type: string
          enum: [login, forgot-password, register]
        pendingToken:
          type: string
      required: [email, otp, purpose]

    VerifyOtpResult:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        resetToken:
          type: string
      required: []

    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]

    ForgotPasswordResult:
      type: object
      properties:
        message:
          type: string
      required: [message]

    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
      required: [token, password]

    ResetPasswordResult:
      type: object
      properties:
        message:
          type: string
      required: [message]

    AuthResult:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, createdAt]

    EncodeInput:
      type: object
      properties:
        file:
          type: string
          format: binary
        message:
          type: string
        key:
          type: string
        algorithm:
          type: string
          enum: [aes-256-gcm, aes-256-cbc, chacha20-poly1305, triple-des]
          default: aes-256-gcm
      required: [file, message]

    EncodeResult:
      type: object
      properties:
        downloadUrl:
          type: string
        filename:
          type: string
        carrier:
          type: string
          enum: [image, audio, video]
        bytesUsed:
          type: integer
        totalBytes:
          type: integer
        timeSec:
          type: number
        algorithmUsed:
          type: ["string", "null"]
      required: [downloadUrl, filename, carrier, bytesUsed, totalBytes, timeSec]

    DecodeInput:
      type: object
      properties:
        file:
          type: string
          format: binary
        key:
          type: string
      required: [file]

    DecodeResult:
      type: object
      properties:
        message:
          type: string
        carrier:
          type: string
          enum: [image, audio, video]
        encrypted:
          type: boolean
        algorithmUsed:
          type: ["string", "null"]
      required: [message, carrier, encrypted]

    DetectInput:
      type: object
      properties:
        file:
          type: string
          format: binary
      required: [file]

    DetectResult:
      type: object
      properties:
        verdict:
          type: string
          enum: [CLEAN, SUSPECT, STEGO, PIXELPEEK]
        probability:
          type: number
        carrier:
          type: string
          enum: [image, audio, video]
        features:
          $ref: "#/components/schemas/StegoFeatures"
      required: [verdict, probability, carrier, features]

    StegoFeatures:
      type: object
      properties:
        entropy:
          type: number
        lsbRatio:
          type: number
        lsbDeviation:
          type: number
        chiSquare:
          type: number
        blockLsbStdev:
          type: number
      required: [entropy, lsbRatio, lsbDeviation, chiSquare, blockLsbStdev]

    Stats:
      type: object
      properties:
        totalOps:
          type: integer
        encodes:
          type: integer
        decodes:
          type: integer
        peeks:
          type: integer
        stegoHits:
          type: integer
      required: [totalOps, encodes, decodes, peeks, stegoHits]

    Event:
      type: object
      properties:
        id:
          type: integer
        operation:
          type: string
          enum: [encode, decode, detect]
        carrier:
          type: string
          enum: [image, audio, video]
        filename:
          type: string
        verdict:
          type: ["string", "null"]
          enum: [CLEAN, SUSPECT, STEGO, PIXELPEEK, null]
        failed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, operation, carrier, filename, failed, createdAt]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]